repository 자료형 long에서 int로 수정

+++ src/main/java/com/smartpotatoo/map_your_trip_be/domain/schedule/dto/ScheduleDateDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 @Data
 public class ScheduleDateDto {
-    private Long schedulesId;
+    private int schedulesId;
     private String date;
     private String content;
     private List<ScheduleTimeDto> times;  // 세부 일정을 위한 리스트
